added twitter auto posts

+++ TarkovBot/Database/Modules/TwitterPost.cs
@@ -4,6 +4,6 @@ public class TwitterPost
 {
     public int Id { get; set; }
     public ulong ChannelId { get; set; }
-    public int AccountId { get; set; }
-    public int? LastTweetId { get; set; }
+    public long AccountId { get; set; }
+    public long? LastTweetId { get; set; }
 }
